fix wiget behaviour
this is a major change in how widgets behave: it introduces
the idea of a focused widget. Before this change, only the
frontmost wiget would receive user input.

Now only the focused widget receives user input. And the
focused widget may also be a parent, in which case the user
input goes directly to the focused widget, even if it may
not be the frontmost (it's children must necessarily be
further up front)

This fixes a number of annoyances, but also brings the
behaviour of ofxWidget further in line with common user
interface conventions.